create test utils, add list beacon blocks
