+++ pynear/src/near/pynear/protos/__init__.py
@@ -1,0 +1,1 @@
+from . import signed_transaction_pb2
